refactor: handle favorites as a new aggregate root
Moving it from another aggregate and creating it's repository
