+++ src/buckstore.products.service.application/buckstore.products.service.application.csproj
@@ -17,6 +17,7 @@
     </ItemGroup>
     
     <ItemGroup>
+      <Folder Include="Commands\Responses" />
       <Folder Include="EventHandlers\Domain" />
     </ItemGroup>
