+++ src/buckstore.products.service.domain/Aggregates/UserFavoriteAggregate/IUserFavoritesRepository.cs
@@ -1,0 +1,9 @@
+﻿using buckstore.products.service.domain.SeedWork;
+
+namespace buckstore.products.service.domain.Aggregates.UserFavoriteAggregate
+{
+    public interface IUserFavoritesRepository : IRepository<ProductFavorites>
+    {
+
+    }
+}
