+++ src/buckstore.products.service.domain/Aggregates/UserFavoriteAggregate/ProductFavorites.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using buckstore.products.service.domain.SeedWork;
 
-namespace buckstore.products.service.domain.Aggregates.ProductAggregate
+namespace buckstore.products.service.domain.Aggregates.UserFavoriteAggregate
 {
-    public class ProductFavorites : Entity
+    public class ProductFavorites : Entity, IAggregateRoot
     {
         public Guid UserId { get; private set; }
         public Guid ProductId { get; private set; }
